LFS-241: Add support for questionnaire sections Move form entry rendering code into its own file
