+++ modules/data-entry/src/main/frontend/src/questionnaire/Form.jsx
@@ -29,21 +29,11 @@ import {
 } from "@material-ui/core";
 
 import QuestionnaireStyle from "./QuestionnaireStyle";
-import AnswerComponentManager from "./AnswerComponentManager";
-import Section from "./Section";
+import FormEntry from "./FormEntry";
 
 // TODO Once components from the login module can be imported, open the login Dialog in-page instead of opening a popup window
 
 // TODO Try to move the save-failed code somewhere more generic instead of the Form component
-
-// FIXME In order for the questions to be registered, they need to be loaded, and the only way to do that at the moment is to explicitly invoke them here. Find a way to automatically load all question types, possibly using self-declaration in a node, like the assets, or even by filtering through assets.
-
-import BooleanQuestion from "./BooleanQuestion";
-import DateQuestion from "./DateQuestion";
-import NumberQuestion from "./NumberQuestion";
-import PedigreeQuestion from "./PedigreeQuestion";
-import TextQuestion from "./TextQuestion";
-import VocabularyQuestion from "./VocabularyQuestion";
 
 /**
  * Component that displays an editable Form.
@@ -135,53 +125,6 @@ function Form (props) {
     setLastSaveStatus(undefined);
   }
 
-  /**
-   * Method responsible for displaying a question from the questionnaire, along with its answer(s).
-   *
-   * @param {Object} questionDefinition the question definition JSON
-   * @param {string} key the node name of the question definition JCR node
-   * @returns a React component that renders the question
-   */
-  let displayQuestion = (questionDefinition, key) => {
-    const existingAnswer = Object.entries(data)
-      .find(([key, value]) => value["sling:resourceSuperType"] == "lfs/Answer"
-        && value["question"]["jcr:uuid"] === questionDefinition["jcr:uuid"]);
-    // This variable must start with an upper case letter so that React treats it as a component
-    const QuestionDisplay = AnswerComponentManager.getAnswerComponent(questionDefinition);
-    return <QuestionDisplay key={key} questionDefinition={questionDefinition} existingAnswer={existingAnswer} />;
-  };
-
-  /**
-   * Method responsible for displaying a section from the questionnaire, along with its answer(s).
-   * TODO: Somehow pass the conditional state upwards from here
-   *
-   * @param {Object} sectionDefinition the section definition JSON
-   * @param {string} key the node name of the section definition JCR node
-   * @returns a React component that renders the section
-   */
-  let displaySection = (sectionDefinition, key) => {
-    // Parse through SectionLink nodes to find our Section node
-    // TODO: What if a sectionRef loop is present?
-    while (sectionDefinition && sectionDefinition["sling:resourceType"] == "sectionLink") {
-      sectionDefinition = sectionDefinition["ref"];
-    }
-    const existingAnswer = Object.entries(data)
-      .find(([key, value]) => value["sling:resourceType"] == "lfs/AnswerSection"
-        && value["question"]["jcr:uuid"] === sectionDefinition["jcr:uuid"]);
-    return <Section key={key} sectionDefinition={sectionDefinition} existingAnswer={existingAnswer} />;
-  }
-
-  // TODO: As before, I'm writing something that's basically an if statement
-  // this should instead be via a componentManager
-  // This code is likely reuseable for the Section component as well
-  let display = (questionDefinition, key) => {
-    if (questionDefinition["jcr:primaryType"] == "lfs:Question") {
-      return displayQuestion(questionDefinition, key);
-    } else if (questionDefinition["jcr:primaryType"] == "lfs:Section") {
-      return displaySection(questionDefinition, key);
-    }
-  }
-
   // If the data has not yet been fetched, return an in-progress symbol
   if (!data) {
     fetchData();
@@ -226,7 +169,7 @@ function Form (props) {
         {
           Object.entries(data.questionnaire)
             .filter(([key, value]) => value['jcr:primaryType'] == 'lfs:Question' || value['jcr:primaryType'] == 'lfs:Section')
-            .map(([key, questionDefinition]) => <Grid item key={key}>{display(questionDefinition, key)}</Grid>)
+            .map(([key, questionDefinition]) => FormEntry(questionDefinition, data, key))
         }
       </Grid>
